+++ libraries/AP_HAL_Empty/RCOutput.h
@@ -4,8 +4,8 @@
 
 class Empty::RCOutput : public AP_HAL::RCOutput {
     void     init() override;
-    void     set_freq(uint32_t chmask, uint16_t freq_hz) override;
-    uint16_t get_freq(uint8_t ch) override;
+    void     set_freq(uint32_t chmask, uint32_t freq_hz) override;
+    uint32_t get_freq(uint8_t ch) override;
     void     enable_ch(uint8_t ch) override;
     void     disable_ch(uint8_t ch) override;
     void     write(uint8_t ch, uint16_t period_us) override;
