+++ libraries/AP_HAL_SITL/RCOutput.h
@@ -9,8 +9,8 @@ class HALSITL::RCOutput : public AP_HAL::RCOutput {
 public:
     explicit RCOutput(SITL_State *sitlState): _sitlState(sitlState), _freq_hz(50) {}
     void init() override;
-    void set_freq(uint32_t chmask, uint16_t freq_hz) override;
-    uint16_t get_freq(uint8_t ch) override;
+    void set_freq(uint32_t chmask, uint32_t freq_hz) override;
+    uint32_t get_freq(uint8_t ch) override;
     void enable_ch(uint8_t ch) override;
     void disable_ch(uint8_t ch) override;
     void write(uint8_t ch, uint16_t period_us) override;
@@ -39,7 +39,7 @@ private:
     SITL_State *_sitlState;
     AP_ESC_Telem_SITL *esc_telem;
 
-    uint16_t _freq_hz;
+    uint32_t _freq_hz;
     uint16_t _enable_mask;
     bool _corked;
     uint16_t _pending[SITL_NUM_CHANNELS];
