+++ libraries/AP_Motors/AP_Motors_Class.h
@@ -226,7 +226,7 @@ protected:
     virtual void        output_armed_stabilizing() = 0;
     virtual void        rc_write(uint8_t chan, uint16_t pwm);
     virtual void        rc_write_angle(uint8_t chan, int16_t angle_cd);
-    virtual void        rc_set_freq(uint32_t mask, uint16_t freq_hz);
+    virtual void        rc_set_freq(uint32_t mask, uint32_t freq_hz);
 
 
     /*
@@ -247,7 +247,7 @@ protected:
 
     // internal variables
     uint16_t            _loop_rate;                 // rate in Hz at which output() function is called (normally 400hz)
-    uint16_t            _speed_hz;                  // speed in hz to send updates to motors
+    uint32_t            _speed_hz;                  // speed in hz to send updates to motors
     float               _roll_in;                   // desired roll control from attitude controllers, -1 ~ +1
     float               _roll_in_ff;                // desired roll feed forward control from attitude controllers, -1 ~ +1
     float               _pitch_in;                  // desired pitch control from attitude controller, -1 ~ +1
